+++ backend/docker.go
@@ -294,11 +294,6 @@ func (p *dockerProvider) Start(ctx gocontext.Context, startAttributes *StartAttr
 
 	logger := context.LoggerFromContext(ctx).WithField("self", "backend/docker_provider")
 
-	cpuSets, err := p.checkoutCPUSets()
-	if err != nil && cpuSets != "" {
-		return nil, err
-	}
-
 	if startAttributes.ImageName != "" {
 		imageName = startAttributes.ImageName
 	} else {
@@ -307,6 +302,7 @@ func (p *dockerProvider) Start(ctx gocontext.Context, startAttributes *StartAttr
 			Infra:    "docker",
 		})
 		if err != nil {
+			logger.WithField("err", err).Error("couldn't select image")
 			return nil, err
 		}
 
@@ -333,7 +329,15 @@ func (p *dockerProvider) Start(ctx gocontext.Context, startAttributes *StartAttr
 		Memory:     int64(p.runMemory),
 		ShmSize:    int64(p.runShm),
 		Tmpfs:      p.tmpFs,
+		CPUSet:     strconv.Itoa(p.runCPUs),
 	}
+
+	cpuSets, err := p.checkoutCPUSets()
+	if err != nil {
+		logger.WithField("err", err).Error("couldn't checkout CPUSets")
+		return nil, err
+	}
+	logger.WithField("cpu_sets", cpuSets).Info("checked out")
 
 	if cpuSets != "" {
 		dockerConfig.CPUSet = cpuSets
@@ -343,14 +347,19 @@ func (p *dockerProvider) Start(ctx gocontext.Context, startAttributes *StartAttr
 	logger.WithFields(logrus.Fields{
 		"config":      fmt.Sprintf("%#v", dockerConfig),
 		"host_config": fmt.Sprintf("%#v", dockerHostConfig),
-	}).Debug("starting container")
+	}).Debug("creating container")
 
+	// FIXME: This doesn't seem to create the container with the Config and HostConfig
 	container, err := p.client.CreateContainer(docker.CreateContainerOptions{
 		Config:     dockerConfig,
 		HostConfig: dockerHostConfig,
 	})
+	container.Config = dockerConfig
+	container.HostConfig = dockerHostConfig
 
 	if err != nil {
+		logger.WithField("err", err).Error("couldn't create container")
+
 		if container != nil {
 			err := p.client.RemoveContainer(docker.RemoveContainerOptions{
 				ID:            container.ID,
@@ -377,6 +386,8 @@ func (p *dockerProvider) Start(ctx gocontext.Context, startAttributes *StartAttr
 	go func(id string) {
 		for {
 			container, err := p.client.InspectContainer(id)
+			container.Config = dockerConfig
+			container.HostConfig = dockerHostConfig
 			if err != nil {
 				errChan <- err
 				return